lots of changes:
new commands - this entire project is currently only supporting txt files (as strings are used to send info to user/remote):
'up' to upload files from a given filepath
'new' to create a new txt file in a given directory
'get' opens a file and allows the user to edit it and view it
'sv' saves the currently opened file
'SELF-DELETE' to delete the folder associated with that user

help messages/warnings now appear in 'virtual' subdirectories e.g. _/fol0/sv/ will be a page displaying help when the user tried to use the 'sv' command in the wrong way in the 'fol0' folder

async server will now read to EOF on incoming messages

Login now works. secure accounts can be created, requiring usernames and passwords to be correctly entered before logging in